+++ eyes.selenium.java/src/test/java/com/applitools/eyes/renderingGrid/TestRenderings.java
@@ -190,7 +190,7 @@ public class TestRenderings extends ReportingTestSuite {
         VisualGridRunner runner = new VisualGridRunner(10);
         Eyes eyes = new Eyes(runner);
         Configuration conf = eyes.getConfiguration();
-        conf.addBrowser(new IosDeviceInfo(IosDeviceName.iPhone_XR, IosScreenOrientation.LANDSCAPE_LEFT));
+        conf.addBrowser(new IosDeviceInfo(IosDeviceName.iPhone_XR, ScreenOrientation.LANDSCAPE));
         conf.setSaveDiffs(false);
         eyes.setConfiguration(conf);
         eyes.setLogHandler(new StdoutLogHandler());